% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test_runner.R
\name{wilcoxon_test_runner}
\alias{wilcoxon_test_runner}
\title{Wilcoxon rank sum test runner}
\usage{
wilcoxon_test_runner(X, G, P = 0.001, NSLOTS = 4)
}
\arguments{
\item{X}{Abundance table (i.e., data.frame or matrix) formatted as samples x features. Sample names should be set as rownames.}

\item{G}{Groups table (i.e., data.frame). This table should have two columns: the first column with the sample names and the second column with groups numbers.}

\item{P}{P-value used to select singificant features. Default 1e-3.}

\item{NSLOTS}{Number of cores used. Default 4.}
}
\value{
This function returns a list where each element corresponds to a group (tested against all the other groups) and consists of a vector of significant features.
}
\description{
This function tests the differential abundance of features (e.g., ASVs, OTUs, or OPUs) between groups using the Wilcoxon rank sum test. The function includes the rarification of the abundance matrix and the filtering of singletons.
}
